+++ packages/enhanced/test/configCases/sharing/layers-consume-loader/test.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  layers: true,
+  findBundle: function () {
+    return ['bundle0.js'];
+  },
+};