+++ packages/error-codes/src/desc.ts
@@ -29,7 +29,7 @@ export const runtimeDescMap = {
 };
 
 export const typeDescMap = {
-  [TYPE_001]: `${failedTo} generate type declaration.`,
+  [TYPE_001]: `${failedTo} to generate type declaration. Execute the below cmd to reproduce and fix the error.`,
 };
 
 export const buildDescMap = {