+++ packages/nextjs-mf/src/plugins/NextFederationPlugin/index.ts
@@ -147,6 +147,27 @@ export class NextFederationPlugin {
       asyncFunction: true,
     };
 
+    // Add layer rules for resource queries
+    if (!compiler.options.module.rules) {
+      compiler.options.module.rules = [];
+    }
+
+    // Add layer rules for RSC, client and SSR
+    compiler.options.module.rules.push({
+      resourceQuery: /\?rsc/,
+      layer: 'rsc',
+    });
+
+    compiler.options.module.rules.push({
+      resourceQuery: /\?client/,
+      layer: 'client',
+    });
+
+    compiler.options.module.rules.push({
+      resourceQuery: /\?ssr/,
+      layer: 'ssr',
+    });
+
     applyPathFixes(compiler, this._options, this._extraOptions);
     if (this._extraOptions.debug) {
       compiler.options.devtool = false;
@@ -181,7 +202,7 @@ export class NextFederationPlugin {
         ...(isServer
           ? [require.resolve('@module-federation/node/runtimePlugin')]
           : []),
-        require.resolve(path.join(__dirname, '../container/runtimePlugin')),
+        require.resolve(path.join(__dirname, '../container/runtimePlugin.cjs')),
         ...(this._options.runtimePlugins || []),
       ].map((plugin) => plugin + '?runtimePlugin'),
       //@ts-ignore
@@ -211,13 +232,7 @@ export class NextFederationPlugin {
   }
 
   private getNoopPath(): string {
-    let noop;
-    try {
-      noop = require.resolve('../../federation-noop');
-    } catch (e) {
-      noop = require.resolve('../../federation-noop.cjs');
-    }
-    return noop;
+    return require.resolve('../../federation-noop.cjs');
   }
 }