+++ packages/sdk/src/types/plugins/ModuleFederationPlugin.ts
@@ -219,7 +219,7 @@ export interface ModuleFederationPluginOptions {
   /**
    * Share scope name used for all shared modules (defaults to 'default').
    */
-  shareScope?: string;
+  shareScope?: string | string[];
   /**
    * load shared strategy(defaults to 'version-first').
    */
@@ -386,7 +386,7 @@ export interface RemotesConfig {
   /**
    * The name of the share scope shared with this remote.
    */
-  shareScope?: string;
+  shareScope?: string | string[];
 }
 /**
  * Modules that should be shared in the share scope. Property names are used to match requested modules in this compilation. Relative requests are resolved, module requests are matched unresolved, absolute paths will match resolved requests. A trailing slash will match all requests with this prefix. In this case shareKey must also have a trailing slash.
@@ -426,7 +426,7 @@ export interface SharedConfig {
   /**
    * Share scope name.
    */
-  shareScope?: string;
+  shareScope?: string | string[];
   /**
    * load shared strategy(defaults to 'version-first').
    */