+++ apps/website-new/CHANGELOG.md
@@ -1,5 +1,11 @@
 # website-new
 
+## 1.0.33
+
+### Patch Changes
+
+- @module-federation/error-codes@0.11.4
+
 ## 1.0.32
 
 ### Patch Changes