+++ packages/managers/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/managers
 
+## 0.11.4
+
+### Patch Changes
+
+- Updated dependencies [64a2bc1]
+- Updated dependencies [c14842f]
+  - @module-federation/sdk@0.11.4
+
 ## 0.11.3
 
 ### Patch Changes