+++ packages/rspack/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @module-federation/rspack
 
+## 0.11.4
+
+### Patch Changes
+
+- Updated dependencies [64a2bc1]
+- Updated dependencies [ed8bda3]
+- Updated dependencies [c14842f]
+  - @module-federation/sdk@0.11.4
+  - @module-federation/dts-plugin@0.11.4
+  - @module-federation/bridge-react-webpack-plugin@0.11.4
+  - @module-federation/managers@0.11.4
+  - @module-federation/manifest@0.11.4
+  - @module-federation/runtime-tools@0.11.4
+  - @module-federation/inject-external-runtime-core-plugin@0.11.4
+
 ## 0.11.3
 
 ### Patch Changes